Crate reconstruction

Modules

operators

tree_repr

Structs

MessagesForSession

SpanId
Describes where a span is located within the trace tree.

Traits

Service
An accessor trait for retrieving the service of a message.

SessionizableMessage
A sessionizable message.

SpanPosition

Functions

canonical_shape

service_calls
Extracts (source, destination) service pairs from a trace tree.

Type Definitions

Degree

Timestamp

TraceId