Function reconstruction::service_calls

pub fn service_calls<M>(messages: &mut Vec<M>) -> Vec<(M::Service, M::Service)> where
    M: SpanPosition + Service, 

Extracts (source, destination) service pairs from a trace tree.

This function returns a list of (transitively) communicating services pair. Any service of a message in the parent span is considered the source service, eventually resulting in invocations at any of the destination services. For example, for three nested servic calls A -> B -> C, the resulting list is [(A, B), (A, C), (B, C)].

Takes a list of messages which implement SpanPosition, meaning the messages must form a hierarchical trace tree encoded in their SpanId.

Note:

This function can mutate the ordering of the messages vector. The current implementation sorts the messages in lexicographical order of the span ids.