Module reconstruction::operators::stats

A collection of various basic statistical Timely operators.

These operators are intended as building blocks for custom operators running on top the stream of reconstructed user-sessions.

Traits

CountNumFragmentsPerSessionPerEpoch
Counts the number of fragmented sessions observed after each epoch.

Histogram
Computes a histogram of the values observed per epoch.

SumPerEpoch
Emits total value (cumulative sum) of a numerical stream once per epoch.

TopK
Calculates the top K most frequent bins per epoch.