Trait reconstruction::operators::stats::Histogram

pub trait Histogram<S: Scope, K, V> {
    type Epoch: Timestamp;
    fn histogram<F>(
        &self, 
        discretizer: F
    ) -> Stream<S, (Self::Epoch, Vec<(V, u64)>)>
    where
        F: Fn(&K) -> V + 'static;
}

Computes a histogram of the values observed per epoch.

This operator takes a custom discretizer function, i.e. a function which extracts a nominal variable (bin) V from a given input K. The frequency within each bin is reported per epoch.

Examples

extern crate timely;
extern crate reconstruction;

use timely::dataflow::operators::{ToStream, Map, Capture};
use timely::dataflow::operators::capture::Extract;
use timely::progress::timestamp::RootTimestamp;

use reconstruction::operators::stats::Histogram;

fn main() {
    let captured = timely::example(|scope| {
        (0..10).to_stream(scope)
           .histogram(|n| n % 2) // bins for `odd` and `even`
           .map_in_place(|x| x.1.sort()) // the created bins are not sorted
           .capture()
    });

    let extracted = captured.extract();
    // half of the elements should be `even` (0), the other half should be `odd` (1)
    assert_eq!(extracted, vec![(RootTimestamp::new(0), vec![(0, vec![(0, 5), (1, 5)])])]);
}

Associated Types

type Epoch: Timestamp

Required Methods

fn histogram<F>(
&self,
discretizer: F
) -> Stream<S, (Self::Epoch, Vec<(V, u64)>)> where
F: Fn(&K) -> V + 'static,

Returns a stream of (timestamp, (bin, frequency))