Trait reconstruction::operators::stats::TopK

pub trait TopK<S: Scope, K, V> {
    type Epoch: Timestamp;
    fn topk<F>(
        &self, 
        discretizer: F, 
        top_k: u64
    ) -> Stream<S, (Self::Epoch, Vec<(V, u64)>)>
    where
        F: Fn(&K) -> V + 'static;
}

Calculates the top K most frequent bins per epoch.

This operator takes a custom discretizer function, i.e. a function which extracts a nominal variable (bin) V from a given input K. It reports the top_k bins of with the highest frequency per epoch.

Examples

extern crate timely;
extern crate reconstruction;

use timely::dataflow::operators::{ToStream, Capture};
use timely::dataflow::operators::capture::Extract;
use timely::progress::timestamp::RootTimestamp;

use reconstruction::operators::stats::TopK;

fn main() {
    let captured = timely::example(|scope| {
        vec![('a', 5), ('b', 1), ('c', 2), ('d', 1)]
           .to_stream(scope)
           .topk(|&(_node, w)| w, 1)
           .capture()
    });

    let extracted = captured.extract();
    // the most common weight is `1` with a frequency of 2
    assert_eq!(extracted, vec![(RootTimestamp::new(0), vec![(0, vec![(1, 2)])])]);
}

Associated Types

type Epoch: Timestamp

Required Methods

fn topk<F>(
&self,
discretizer: F,
top_k: u64
) -> Stream<S, (Self::Epoch, Vec<(V, u64)>)> where
F: Fn(&K) -> V + 'static,

Returns a stream of (timestamp, (bin, frequency))